refactor!: shrink block heights to u32 across the stack
Bitcoin height fits a u32 for ~81,000 years — Core itself stores it as
a 32-bit int. Converts once at the bitcoind RPC boundary and carries
u32 through the block count vote, the LN contract expiries, and the
wallet tx chain stamps.

+++ picomint-bitcoin-rpc/src/lib.rs
@@ -30,7 +30,7 @@ pub struct Feerate {
 #[derive(Debug, Clone)]
 pub struct BitcoinRpcStatus {
     pub network: Network,
-    pub block_count: u64,
+    pub block_count: u32,
     /// `None` while the backend is still syncing — fee estimation has no
     /// data until the node is at the tip, and consensus (the only consumer
     /// that needs a feerate) doesn't start until then either.
@@ -113,7 +113,7 @@ impl BitcoinRpcMonitor {
         self.rpc.get_block(hash).await
     }
 
-    pub async fn get_block_hash(&self, height: u64) -> Result<BlockHash> {
+    pub async fn get_block_hash(&self, height: u32) -> Result<BlockHash> {
         ensure!(
             self.status_rx.borrow().is_some(),
             "Not connected to bitcoin backend"
